Move main resolver to delegation resolver in order to use field resolvers
